+++ macros/src/attr/variant.rs
@@ -1,6 +1,6 @@
-use syn::{Attribute, Ident, Result, Type};
+use syn::{Attribute, Fields, Ident, Result, Type, Variant};
 
-use super::EnumAttr;
+use super::{Attr, Serde};
 use crate::{
     attr::{parse_assign_from_str, parse_assign_inflection, parse_assign_str, Inflection},
     utils::parse_attrs,
@@ -17,42 +17,74 @@ pub struct VariantAttr {
     pub untagged: bool,
 }
 
-#[cfg(feature = "serde-compat")]
-#[derive(Default)]
-pub struct SerdeVariantAttr(VariantAttr);
-
 impl VariantAttr {
-    pub fn new(attrs: &[Attribute], enum_attr: &EnumAttr) -> Result<Self> {
-        let mut result = Self::default();
-        parse_attrs(attrs)?.for_each(|a| result.merge(a));
-        result.rename_all = result.rename_all.or(enum_attr.rename_all_fields);
+    pub fn from_attrs(attrs: &[Attribute]) -> Result<Self> {
+        let mut result = parse_attrs::<Self>(attrs)?;
         #[cfg(feature = "serde-compat")]
         if !result.skip {
-            crate::utils::parse_serde_attrs::<SerdeVariantAttr>(attrs)
-                .for_each(|a| result.merge(a.0));
+            let serde_attr = crate::utils::parse_serde_attrs::<VariantAttr>(attrs);
+            result = result.merge(serde_attr.0);
         }
         Ok(result)
     }
+}
 
-    fn merge(
-        &mut self,
-        VariantAttr {
-            type_as,
-            type_override,
-            rename,
-            rename_all,
-            inline,
-            skip,
-            untagged,
-        }: VariantAttr,
-    ) {
-        self.type_as = self.type_as.take().or(type_as);
-        self.type_override = self.type_override.take().or(type_override);
-        self.rename = self.rename.take().or(rename);
-        self.rename_all = self.rename_all.take().or(rename_all);
-        self.inline = self.inline || inline;
-        self.skip = self.skip || skip;
-        self.untagged = self.untagged || untagged;
+impl Attr for VariantAttr {
+    type Item = Variant;
+
+    fn merge(self, other: Self) -> Self {
+        Self {
+            type_as: self.type_as.or(other.type_as),
+            type_override: self.type_override.or(other.type_override),
+            rename: self.rename.or(other.rename),
+            rename_all: self.rename_all.or(other.rename_all),
+            inline: self.inline || other.inline,
+            skip: self.skip || other.skip,
+            untagged: self.untagged || other.untagged,
+        }
+    }
+
+    fn assert_validity(&self, item: &Self::Item) -> Result<()> {
+        if self.type_as.is_some() {
+            if self.type_override.is_some() {
+                syn_err_spanned!(
+                    item;
+                    "`as` is not compatible with `type`"
+                )
+            }
+
+            if self.rename_all.is_some() {
+                syn_err_spanned!(
+                    item;
+                    "`as` is not compatible with `rename_all`"
+                )
+            }
+        }
+
+        if self.type_override.is_some() {
+            if self.rename_all.is_some() {
+                syn_err_spanned!(
+                    item;
+                    "`type` is not compatible with `rename_all`"
+                )
+            }
+
+            if self.inline {
+                syn_err_spanned!(
+                    item;
+                    "`type` is not compatible with `inline`"
+                )
+            }
+        }
+      
+        if !matches!(item.fields, Fields::Named(_)) && self.rename_all.is_some() {
+            syn_err_spanned!(
+                item;
+                "`rename_all` is not applicable to unit or tuple variants"
+            )
+        }
+
+        Ok(())
     }
 }
 
@@ -70,7 +102,7 @@ impl_parse! {
 
 #[cfg(feature = "serde-compat")]
 impl_parse! {
-    SerdeVariantAttr(input, out) {
+    Serde<VariantAttr>(input, out) {
         "rename" => out.0.rename = Some(parse_assign_str(input)?),
         "rename_all" => out.0.rename_all = Some(parse_assign_inflection(input)?),
         "skip" => out.0.skip = true,